requirements.txt
Django==4.1.3
djangorestframework==3.14.0
joblib==1.3.2
python-dotenv==1.0.0
spacy==3.7.4
scikit-learn==1.6.1
numpy==1.26.4
pandas==2.2.2
gunicorn==21.2.0
nltk==3.8.1
psycopg2-binary==2.9.9